Time: 0 ms (100%), Space: 131.4 MB (93.32%) - LeetHub

+++ 2962-count-subarrays-where-max-element-appears-at-least-k-times/2962-count-subarrays-where-max-element-appears-at-least-k-times.cpp
@@ -5,7 +5,7 @@ public:
         int left =0 ,right = 0;
         long long count = 0;
         int MaxVal = nums[0];
-        long long freq = 0;
+        int freq = 0;
         for(int x : nums){
             MaxVal = max(MaxVal , x);
         }
